+++ packages/clients/textLocator/src/utils/coastalGazetteer/searchToponym.ts
@@ -1,7 +1,11 @@
 import { FeatureCollection } from 'geojson'
 import { Map } from 'ol'
 import { PolarStore, SelectResultFunction } from '@polar/lib-custom-types'
-import { SearchResultSymbols } from '@polar/plugin-address-search'
+import {
+  SearchResultSymbols,
+  AddressSearchGetters,
+  AddressSearchState,
+} from '@polar/plugin-address-search'
 import VectorSource from 'ol/source/Vector'
 import {
   GeometrySearchGetters,
@@ -51,10 +55,10 @@ export async function searchCoastalGazetteerByToponym(
   )
 }
 
-export const selectResult: SelectResultFunction = (
-  { commit, rootGetters },
-  { feature }
-) => {
+export const selectResult: SelectResultFunction<
+  AddressSearchState,
+  AddressSearchGetters
+> = ({ commit, rootGetters }, { feature }) => {
   // default behaviour (AddressSearch selects and is not involved in further behaviour)
   commit('plugin/addressSearch/setChosenAddress', feature, { root: true })
   commit('plugin/addressSearch/setInputValue', feature.title, { root: true })